tests/requirements.txt
pytest>=6.2.5
pytest-django>=4.5.2
coverage>=6.2
factory-boy>=3.2.1